refactor(DatagridSlug): implement types
